pyproject.toml
[tool.poetry]
name = "efmarketplace"
version = "0.1.0"
description = ""
authors = [

]
maintainers = [

]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.85.0"
uvicorn = { version = "^0.18.3", extras = ["standard"] }
pydantic = {version = "^1.10.2", extras = ["dotenv"]}
yarl = "^1.8.1"
ujson = "^5.5.0"
tortoise-orm = "^0.19.2"
asyncpg = "^0.26.0"
aiofiles = "^22.1.0"
httptools = "^0.5.0"
loguru = "^0.6.0"
bcrypt = "^4.0.1"
fastapi-jwt = "^0.1.8"
fastapi-jwt-auth = "^0.5.0"
captcha = "^0.4"
redis = "^4.5.0"
httpx = "^0.23.3"
fastapi-utils = "^0.2.1"
aerich = "^0.7.1"
fakeredis = "^2.9.0"

[tool.poetry.dev-dependencies]
pytest = "^7.1.3"
flake8 = "~4.0.1"
mypy = "^0.981"
isort = "^5.10.1"
pre-commit = "^2.20.0"
wemake-python-styleguide = "^0.17.0"
black = "^22.8.0"
autoflake = "^1.6.1"
pytest-cov = "^4.0.0"
anyio = "^3.6.1"
pytest-env = "^0.6.2"
asynctest = "^0.13.0"
nest-asyncio = "^1.5.6"
httpx = "^0.23.0"

[tool.isort]
profile = "black"
multi_line_output = 3
src_paths = ["efmarketplace",]

[tool.mypy]
strict = true
ignore_missing_imports = true
allow_subclassing_any = true
allow_untyped_calls = true
pretty = true
show_error_codes = true
implicit_reexport = true
allow_untyped_decorators = true
warn_unused_ignores = false
warn_return_any = false
namespace_packages = true

# Remove this and add `types-redis`
# when the issue https://github.com/python/typeshed/issues/8242 is resolved.
[[tool.mypy.overrides]]
module = [
    'redis.asyncio'
]
ignore_missing_imports = true

[tool.pytest.ini_options]
filterwarnings = [
    "error",
    "ignore::DeprecationWarning",
    "ignore:.*unclosed.*:ResourceWarning",
]
env = [
    "EFMARKETPLACE_DB_BASE=efmarketplace_test",
]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.aerich]
tortoise_orm = "efmarketplace.db.config.TORTOISE_CONFIG"
location = "./efmarketplace/db/migrations"
src_folder = "./efmarketplace/db"
